Remove frontend, add telegram bot dep
