+++ .gitignore
@@ -23,3 +23,30 @@ yarn-debug.log*
 yarn-error.log*
 
 *.code-workspace
+
+# Logs
+logs
+*.log
+
+# Transpiled files
+build/
+
+# VS Code
+.vscode
+!.vscode/tasks.js
+
+# JetBrains IDEs
+.idea/
+
+# Optional npm cache directory
+.npm
+
+# Optional eslint cache
+.eslintcache
+
+# Misc
+.DS_Store
+
+# Environment variables file
+.env
+
