--- backend/.gitignore
@@ -1,32 +1,0 @@
-# Logs
-logs
-*.log
-npm-debug.log*
-
-# Dependencies
-node_modules/
-
-# Coverage
-coverage
-
-# Transpiled files
-build/
-
-# VS Code
-.vscode
-!.vscode/tasks.js
-
-# JetBrains IDEs
-.idea/
-
-# Optional npm cache directory
-.npm
-
-# Optional eslint cache
-.eslintcache
-
-# Misc
-.DS_Store
-
-# Environment variables file
-.env
