--- frontend/.dockerignore
@@ -1,18 +1,0 @@
-node_modules/
-build/
-
-Dockerfile
-.dockerignore
-
-.env.example
-
-.github/
-.git/
-.gitignore
-
-.DS_Store
-*/.DS_Store
-
-yarn-error.log
-
-taskdef.json
