--- frontend/.eslintrc.js
@@ -1,41 +1,0 @@
-module.exports = {
-  parser: 'babel-eslint',
-
-  extends: ['airbnb', 'prettier', 'prettier/react'],
-
-  plugins: ['prettier'],
-
-  env: {
-    browser: true,
-  },
-
-  rules: {
-    'prettier/prettier': 'error',
-    'react/jsx-filename-extension': [
-      'error',
-      { extensions: ['.js', '.ts', '.jsx', '.tsx'] },
-    ],
-    'import/extensions': [
-      'error',
-      'ignorePackages',
-      {
-        js: 'never',
-        jsx: 'never',
-        ts: 'never',
-        tsx: 'never',
-      },
-    ],
-    'no-console': ['error', { allow: ['info', 'warn', 'error'] }],
-  },
-
-  settings: {
-    // Allow absolute paths in imports, e.g. import Button from 'components/Button'
-    // https://github.com/benmosher/eslint-plugin-import/tree/master/resolvers
-    'import/resolver': {
-      node: {
-        moduleDirectory: ['node_modules', 'src'],
-        extensions: ['.js', '.jsx', '.ts', '.tsx', '.json'],
-      },
-    },
-  },
-}
