--- frontend/.prettierrc.js
@@ -1,14 +1,0 @@
-module.exports = {
-  printWidth: 80,
-  singleQuote: true,
-  trailingComma: 'all',
-  semi: false,
-  overrides: [
-    {
-      files: "*.ts",
-      options: {
-        parser: "typescript"
-      }
-    }
-  ]
-}
