--- frontend/public/css/util.css
@@ -1,3 +1,0 @@
-.text-center {
-  text-align: center;
-}
