+++ package.json
@@ -21,6 +21,7 @@
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "node-telegram-bot-api": "^0.57.0",
     "tslib": "^2.1.0"
   },
   "devDependencies": {
